Add ChaCha20 DRNG self test
In addition to the ChaCha20 block self test, add a ChaCha20 DRNG self
test covering:

- one block to be generated by the DRNG

- two blocks to be generated by the DRNG with two block reseed

- one block and one byte to be generated by the DRNG with one block and
one byte reseed

The self test is invoked during initialization time.

The self test updates the DRNG state before the state is reseeded. It is
considered harmless to not clear the ChaCha20 DRNG state after the self
test because it contains no entropy to begin with. Before the patch, the
state before the first seeding was zero, now it is the state left by the
self test.

The patch also contains a bug fix in drng_chacha20_alloc: if mlock fails,
deallocate memory.

Signed-off-by: Stephan Mueller <smueller@chronox.de>