fix increment of nonce
The nonce is a 96 bit integer that is intended to be incremented as like
a regular integer. The code, however, performs an increment of all three
nonce words if all of them are zero. Thus, the effective nonce size is
only 94 bits. Albeit this does not affect the security of the DRNG
considering that there is a reseed way before reaching 2^94 generate
operations (where the current nonce would wrap), it is an uncleanliness
which needs to be fixed.

Reported-by: Aaron Voisina
Signed-off-by: Stephan Mueller <smueller@chronox.de>